// src/model/stripeTransactionStatus.ts
export type StripeTransactionStatus =
  | 'INTENT_CREATE_SUCCEEDED'
  | 'INTENT_REQUIRES_PAYMENT_METHOD'
  | 'INTENT_REQUIRES_CONFIRMATION'
  | 'INTENT_REQUIRES_ACTION'
  | 'INTENT_PROCESSING'
  | 'INTENT_REQUIRES_CAPTURE'
  | 'INTENT_SUCCEEDED'
  | 'INTENT_FAILED'
  | 'INTENT_CANCELLED'
  | 'TRANSFER_CREATE_SUCCEEDED'
  | 'TRANSFER_REVERSED_SUCCEEDED'
  | 'TRANSFER_FAILED'
  | 'CHARGE_SUCCEEDED'
  | 'CHARGE_UPDATE_SUCCEEDED'
  | 'CHARGE_FAILED'
  | 'CHARGE_REFUND_SUCCEEDED'
  | 'REFUND_CREATE_SUCCEEDED'
  | 'REFUND_UPDATE_SUCCEEDED'
  | 'REFUND_CHARGE_UPDATE_SUCCEEDED'
  | 'REFUND_FAILED';